+++ test/ch/connection_test.exs
@@ -140,13 +140,16 @@ defmodule Ch.ConnectionTest do
 
     assert Ch.query!(conn, "select {$0:DateTime64(6,'Europe/Moscow')} as d, toString(d)", [utc]).rows ==
              [[msk, "2021-01-01 15:00:00.123456"]]
+  end
 
-    # this test case gaurds against a previous bug where DateTimes with a microsecond value of 0 and precision > 0 would
-    # get encoded as a val like "1.6095024e9" which Clickhouse would be unable to parse to a DateTime.
-    utc_with_zero_microsec = ~U[2021-01-01 12:00:00.000000Z]
-    naive_with_zero_microsec = utc_with_zero_microsec |> DateTime.shift_zone!(Ch.Test.clickhouse_tz(conn)) |> DateTime.to_naive()
-    assert Ch.query!(conn, "select {$0:DateTime64(6)} as d, toString(d)", [utc_with_zero_microsec]).rows ==
-             [[~N[2021-01-01 12:00:00.000000], to_string(naive_with_zero_microsec)]]
+  test "utc datetime64 zero microseconds query param encoding", %{conn: conn} do
+    # this test case guards against a previous bug where DateTimes with a microsecond value of 0 and precision > 0 would
+    # get encoded as a val like "1.6095024e9" which ClickHouse would be unable to parse to a DateTime.
+    utc = ~U[2021-01-01 12:00:00.000000Z]
+    naive = DateTime.to_naive(utc)
+
+    assert Ch.query!(conn, "select {$0:DateTime64(6)} as d, toString(d)", [utc]).rows ==
+             [[~N[2021-01-01 12:00:00.000000], to_string(naive)]]
   end
 
   test "select with options", %{conn: conn} do